\encoding{UTF-8}
\name{carabid}
\alias{carabid}
\docType{data}
\title{Pitfall data on Carabids (Siilinjärvi, Finland) }
\usage{data(carabid)}
\description{
The \code{carabid} data frame has 64 rows representing sampling sites
(pitfalls)  and 45 columns giving counts of Carabid species.
}
\format{
  This data frame contains pitfall counts on 45 Carabid species with
  abbreviated names.  
}
\source{
  The data were collected by Dr. Jarmo K. Holopainen (University of
  Kuopio). 
}
\note{
  These are pitfall data: so they are counts with no absolute upper
  limit, and the total count varies much among observations.
  }
\references{
  Holopainen, J.K., Bergman, T., Hautala, E.-L. and Oksanen,
  J. (1995). The ground beetle fauna (Coleoptera: Carabidae) in relation
  to soil properties and foliar fluoride content in spring
  cereals. Pedobiologia 39, 193--206.
}
\examples{
data(carabid)
}
\keyword{datasets}
